requires 5.8.1
There's actually no requirement in this module that requires 5.8,
however: a) the build tool (Minilla) does b) what this module provides
can be easily replicated with `eval "use Module; 1" or plan skip_all => ...`
and c) This module requires Test::More 0.61 or later, which is core
since perl 5.8.8.

If one wants to support 5.6, requiring this module to require a higher
version of Test::More sounds something they want to avoid altogether.